win32: interpret root path /X/... as windows X:/...
Where X is alpha. This means that if there's a single-letter dir X at
the root of the current drive, it's no longer possible to reference it
as /X/...  of the awkward "dir X at the root of the current drive"
(however, cd /X - without following '/' would work)

It's unclear whether there are remaining cases, or too-many, or other
issues, but these are tested (briefly) and seem to work as expected:

  cd, rm, ls, test -[exd], path-expansion (glob), tab completion,
  exec (and normal commands)

Known issues: when invoking a program, while /X/... is converted to
X:/... for CreateProcess (et al), argv[0] of the called program
remains /X/... so a normal win32 app can't rely on argv[0] resolving
to a valid path to self.