win32: fake drive-letter root path: support /X too
Previously, only "/X/" [prefix] was converted, while a string of
exactly "/X" was not modified at all (for any alpha X).

Now /X alone is converted to "X:/", same as /X/ would.

Known issues: /c* glob doesn't work, /c<tab> doesn't complete
(but /c/* does work, and also /c/<tab>).

Unknown issues: there could be cases where /X and /X/ should differ
but we end up handling them the same (but hopefully only low level).

Details:

Previously we had an inconsistenty because "/c/" and "/c"
referred to different locations (the former is "c:/", and the latter
was, like before, dir "c" at the root of the current drive - which
depends on which drive CWD is at).

Whether or not making /X and /X/ equivalent has negative impact is
yet unknown, because a path ending in "/" is not always interpreted
the same as the same path without trailing /, even if both paths
still have "/" in them (and they do - both begin with '/').

However we do try to make the conversion as close to the win32
API as possible, so that the converted strings hopefully don't keep
lingering too much. But there could still be a negative impact.

But, test -d/-e for /X and /X/ do seem to work the same (both with
and without this commit), and also 'ls' and 'cd' seem to not be
affected by a trailing slash (again, with or without this commit).

Because now there's a case where 2-chars paths become 3-chars, the
inplace variant now requires at least 4 bytes siz (3 chars + \0).
We enforce it with mk[s]temp, and elsewhere checked the call sites.