win32: UTF8_NATIVE part 1: argv/env/spawn
Here we set the framefork for native unicode support, which is
an alternative to UTF8_MANIFEST, but works on windows XP or later.

It unifies the "is utf8" test of these utf8 modes, and makes them
interchangable as far as other unicode features go - like utf8
input/output, unicode editing at the shell prompt, etc.

For now, it supports enabling both native and manifest utf8, but
ultimately they should be mutually exclusive.

Beyond the framework, it converts argv and environ to UTF8 on entry
(from unicode), and when spawning, converts from UTF8 for the W API.
Between entry and spawn, putenv/getenv/etc also work correctly
(because the ANSI env is now UTF8) and reflected correctly at spawn.

It is, however, partial, because except for spawn(ve), all the other
APIs remain the ANSI variants, and so e.g. fopen("<utf8 name>",...)
does not yet work correctly.

The next step would be to write UTF8 variants of other win32 APIs,
and then map them from the ANSI variants, similar to what the UTF8
manifest does, but at build time.