(local) ash: tab-complete variable names
This is inefficient in O(N^2) - just like functions/aliases completion,
but as long as we don't have thousands of var names, it's hopefully ok.

(Tested 10000 names (10K), i7-8750H capped at 2GHz: less than one sec).

It's possible to use a better API where the shell adds all the matches
in one pass, which would be O(N), but for now not worth the effort.